src/python/services/queue-service/requirements.txt
azure-identity==1.4.0b5
azure-storage-queue==12.1.2
azure-ai-formrecognizer==1.0.0b3
azure-ai-textanalytics==1.0.0
azure-cosmos==4.0.0
python-dotenv==0.13.0
dotmap==1.3.17